fixed the structure of the gif creations. Some gifs were not creating well, now they work good

+++ gif_creation/gif_creation_of_algorithms/quick_sort_naive_gif.py
@@ -69,9 +69,11 @@ class SortingVisualizer:
                 writer.append_data(image)
 
         for filename in filenames:
-            os.remove(filename)
-        os.rmdir(frame_dir)
+            if os.path.exists(filename):  
+                os.remove(filename)
+        if os.path.exists(frame_dir):  
+            os.rmdir(frame_dir)
 
 def run(lst):
     visualizer = SortingVisualizer(lst)
-    visualizer.quick_sort_naive_gif(duration=0.5)  
+    visualizer.quick_sort_naive_gif(duration=10)
